templates/requirements.txt
Flask==2.1.3
textblob==0.17.1
nltk==3.8.1
googlesearch-python==1.0.1
